+++ contracts/protocols/perps/pendle/interfaces/Camelot/ICamelotFactory.sol
@@ -18,5 +18,5 @@ interface ICamelotFactory {
     function createPair(address tokenA, address tokenB) external returns (address pair);
 
     function setFeeTo(address) external;
-    function feeInfo() external view returns (uint _ownerFeeShare, address _feeTo);
+    function feeInfo() external view returns (uint256 _ownerFeeShare, address _feeTo);
 }
