+++ contracts/protocols/perps/pendle/interfaces/Thena/IThenaFactory.sol
@@ -3,11 +3,11 @@
 pragma solidity ^0.8.0;
 
 interface IThenaFactory {
-    function allPairsLength() external view returns (uint);
+    function allPairsLength() external view returns (uint256);
 
     function isPair(address pair) external view returns (bool);
 
-    function allPairs(uint index) external view returns (address);
+    function allPairs(uint256 index) external view returns (address);
 
     function pairCodeHash() external pure returns (bytes32);
