+++ contracts/protocols/tokens/stable/frax/ERC20/ERC20PermissionedMint.sol
@@ -9,7 +9,6 @@ import "@crane/contracts/external/openzeppelin-contracts/access/AccessControl.so
 import "@crane/contracts/protocols/tokens/stable/frax/Staking/Owned.sol";
 
 contract ERC20PermissionedMint is ERC20, ERC20Burnable, Owned {
-
     // Core
     address public timelock_address;
 
@@ -19,18 +18,12 @@ contract ERC20PermissionedMint is ERC20, ERC20Burnable, Owned {
 
     /* ========== CONSTRUCTOR ========== */
 
-    constructor(
-        address _creator_address,
-        address _timelock_address,
-        string memory _name,
-        string memory _symbol
-    ) 
-    ERC20(_name, _symbol) 
-    Owned(_creator_address)
+    constructor(address _creator_address, address _timelock_address, string memory _name, string memory _symbol)
+        ERC20(_name, _symbol)
+        Owned(_creator_address)
     {
-      timelock_address = _timelock_address;
+        timelock_address = _timelock_address;
     }
-
 
     /* ========== MODIFIERS ========== */
 
@@ -40,9 +33,9 @@ contract ERC20PermissionedMint is ERC20, ERC20Burnable, Owned {
     }
 
     modifier onlyMinters() {
-       require(minters[msg.sender] == true, "Only minters");
+        require(minters[msg.sender] == true, "Only minters");
         _;
-    } 
+    }
 
     /* ========== RESTRICTED FUNCTIONS ========== */
 
@@ -52,33 +45,33 @@ contract ERC20PermissionedMint is ERC20, ERC20Burnable, Owned {
         emit TokenMinterBurned(b_address, msg.sender, b_amount);
     }
 
-    // This function is what other minters will call to mint new tokens 
+    // This function is what other minters will call to mint new tokens
     function minter_mint(address m_address, uint256 m_amount) public onlyMinters {
         super._mint(m_address, m_amount);
         emit TokenMinterMinted(msg.sender, m_address, m_amount);
     }
 
-    // Adds whitelisted minters 
+    // Adds whitelisted minters
     function addMinter(address minter_address) public onlyByOwnGov {
         require(minter_address != address(0), "Zero address detected");
 
         require(minters[minter_address] == false, "Address already exists");
-        minters[minter_address] = true; 
+        minters[minter_address] = true;
         minters_array.push(minter_address);
 
         emit MinterAdded(minter_address);
     }
 
-    // Remove a minter 
+    // Remove a minter
     function removeMinter(address minter_address) public onlyByOwnGov {
         require(minter_address != address(0), "Zero address detected");
         require(minters[minter_address] == true, "Address nonexistent");
-        
+
         // Delete from the mapping
         delete minters[minter_address];
 
         // 'Delete' from the array by setting the address to 0x0
-        for (uint i = 0; i < minters_array.length; i++){ 
+        for (uint256 i = 0; i < minters_array.length; i++) {
             if (minters_array[i] == minter_address) {
                 minters_array[i] = address(0); // This will leave a null in the array and keep the indices the same
                 break;
@@ -89,7 +82,7 @@ contract ERC20PermissionedMint is ERC20, ERC20Burnable, Owned {
     }
 
     /* ========== EVENTS ========== */
-    
+
     event TokenMinterBurned(address indexed from, address indexed to, uint256 amount);
     event TokenMinterMinted(address indexed from, address indexed to, uint256 amount);
     event MinterAdded(address minter_address);
