+++ contracts/protocols/tokens/stable/frax/ERC20/wfrxETH.sol
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: GPL-3.0
 pragma solidity ^0.8.35;
 
-import {BetterEfficientHashLib} from '@crane/contracts/utils/BetterEfficientHashLib.sol';
-import {BetterAddress} from '@crane/contracts/utils/BetterAddress.sol';
+import {BetterEfficientHashLib} from "@crane/contracts/utils/BetterEfficientHashLib.sol";
+import {BetterAddress} from "@crane/contracts/utils/BetterAddress.sol";
 import "@crane/contracts/external/openzeppelin-contracts/token/ERC20/extensions/IERC20Permit.sol";
 import "@crane/contracts/external/openzeppelin-contracts/utils/cryptography/ECDSA.sol";
 import "@crane/contracts/external/openzeppelin-contracts/utils/cryptography/EIP712.sol";
@@ -31,7 +31,6 @@ import "./ERC20Permit/Counters.sol";
 // Dennis: https://github.com/denett
 // Drake Evans: https://github.com/DrakeEvans
 // Justin Moore: https://github.com/0xJM
-
 
 // !!!!!!! CREDIT GOES TO DAPPHUB AND OPENZEPPELIN !!!!!!!
 
@@ -75,18 +74,17 @@ contract wfrxETH is IERC20Permit, EIP712 {
     // WETH9 State Variables
     // =============================================
 
-    string public name     = "Wrapped Frax Ether";
-    string public symbol   = "wfrxETH";
-    uint8  public decimals = 18;
+    string public name = "Wrapped Frax Ether";
+    string public symbol = "wfrxETH";
+    uint8 public decimals = 18;
 
-    event  Approval(address indexed src, address indexed guy, uint wad);
-    event  Transfer(address indexed src, address indexed dst, uint wad);
-    event  Deposit(address indexed dst, uint wad);
-    event  Withdrawal(address indexed src, uint wad);
+    event Approval(address indexed src, address indexed guy, uint256 wad);
+    event Transfer(address indexed src, address indexed dst, uint256 wad);
+    event Deposit(address indexed dst, uint256 wad);
+    event Withdrawal(address indexed src, uint256 wad);
 
-    mapping (address => uint)                       public  balanceOf;
-    mapping (address => mapping (address => uint))  public  allowance;
-
+    mapping(address => uint256) public balanceOf;
+    mapping(address => mapping(address => uint256)) public allowance;
 
     // Constructor
     // =============================================
@@ -109,7 +107,7 @@ contract wfrxETH is IERC20Permit, EIP712 {
         emit Deposit(msg.sender, msg.value);
     }
 
-    function withdraw(uint wad) public {
+    function withdraw(uint256 wad) public {
         require(balanceOf[msg.sender] >= wad);
         balanceOf[msg.sender] -= wad;
         // payable(msg.sender).transfer(wad);
@@ -117,24 +115,21 @@ contract wfrxETH is IERC20Permit, EIP712 {
         emit Withdrawal(msg.sender, wad);
     }
 
-    function totalSupply() public view returns (uint) {
+    function totalSupply() public view returns (uint256) {
         return address(this).balance;
     }
 
-    function approve(address guy, uint wad) public returns (bool) {
+    function approve(address guy, uint256 wad) public returns (bool) {
         allowance[msg.sender][guy] = wad;
         emit Approval(msg.sender, guy, wad);
         return true;
     }
 
-    function transfer(address dst, uint wad) public returns (bool) {
+    function transfer(address dst, uint256 wad) public returns (bool) {
         return transferFrom(msg.sender, dst, wad);
     }
 
-    function transferFrom(address src, address dst, uint wad)
-        public
-        returns (bool)
-    {
+    function transferFrom(address src, address dst, uint256 wad) public returns (bool) {
         require(balanceOf[src] >= wad);
 
         if (src != msg.sender && allowance[src][msg.sender] != type(uint256).max) {
@@ -150,22 +145,17 @@ contract wfrxETH is IERC20Permit, EIP712 {
         return true;
     }
 
-
     // EIP-2612 Functions
     // =============================================
 
     /**
      * @dev See {IERC20Permit-permit}.
      */
-    function permit(
-        address owner,
-        address spender,
-        uint256 value,
-        uint256 deadline,
-        uint8 v,
-        bytes32 r,
-        bytes32 s
-    ) public virtual override {
+    function permit(address owner, address spender, uint256 value, uint256 deadline, uint8 v, bytes32 r, bytes32 s)
+        public
+        virtual
+        override
+    {
         require(block.timestamp <= deadline, "ERC20Permit: expired deadline");
 
         // For WETH9, msg.sender is always the "owner". Leaving as a function parameter to not break existing ABIs.
@@ -208,7 +198,6 @@ contract wfrxETH is IERC20Permit, EIP712 {
         nonce.increment();
     }
 }
-
 
 /*
                     GNU GENERAL PUBLIC LICENSE
